Completed the shift to full modular * Removed a bunch of `resources.lib.` in favour of the relative `.` * Removed a bunch of Globals/Settings static instantiation * Moved previously missing functions into common, network and playback modules * Moved the remainder of `default.py` inside `resources/lib/startup.py:EntryPoint()` * Moved the initialisation of AmazonTLD/PrimeVideo into common's `InitialiseProvider` * Moved the bootstrapping of AmazonTLD/PrimeVideo into the respective `__init__`s * Given the nature of the `getConfig`/`writeConfig`, the static attribute in those functions are initialised inside `Globals.__init__` * Minor fixes
