+++ plugin.video.amazon-test/resources/lib/l10n.py
@@ -4,13 +4,11 @@ import xbmc
 
 
 def getString(string_id, addonInstance=None):
-    from resources.lib.common import Globals
     if string_id < 30000:
         src = xbmc
     elif addonInstance is None:
-        if not hasattr(getString, 'g'):
-            getString.g = Globals()
-        src = getString.g.addon
+        from .common import Globals
+        src = Globals().addon
     else:
         src = addonInstance
     locString = src.getLocalizedString(string_id)
